requirements.txt
annotated-types==0.7.0
cffi==1.17.1
cryptography==43.0.3
ecdsa==0.19.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyOpenSSL==24.2.1
six==1.16.0
typing_extensions==4.12.2
